Create a new class for LightAdminRoles and LightAdminPrivileges tests, add to it the importImage... method, start using this class method in LightAdminRoles tests
